+++ src/keys/ed25519.js
@@ -16,24 +16,20 @@ exports.generateKey = function (callback) {
   try {
     keys = nacl.sign.keyPair()
   } catch (err) {
-    done(err)
-    return
+    return done(err)
   }
   done(null, keys)
 }
 
 // seed should be a 32 byte uint8array
 exports.generateKeyFromSeed = function (seed, callback) {
-  const done = (err, res) => setImmediate(() => {
-    callback(err, res)
-  })
+  const done = (err, res) => setImmediate(() => callback(err, res))
 
   let keys
   try {
     keys = nacl.sign.keyPair.fromSeed(seed)
   } catch (err) {
-    done(err)
-    return
+    return done(err)
   }
   done(null, keys)
 }
